core/requirements.txt
astroid==2.13.2
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
dill==0.3.6
et-xmlfile==1.1.0
flake8==6.0.0
idna==3.4
isort==5.11.4
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
openpyxl==3.0.10
platformdirs==2.6.2
pycodestyle==2.10.0
pyflakes==3.0.1
pylint==2.15.9
python-decouple==3.6
requests==2.28.1
tomli==2.0.1
tomlkit==0.11.6
types-requests==2.28.11.7
types-urllib3==1.26.25.4
typing_extensions==4.4.0
urllib3==1.26.13
wrapt==1.14.1
